+++ CHANGELOG.md
@@ -6,6 +6,14 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ### Changed
+- [Core] Change `<SearchInput>` behavior:
+    - Can be controlled now, via props `value`, `onChange` and `onReset`.
+    - No longer trigger `onSearch` when input blur by default. You can enable this behavior by setting prop `searchOnInputBlur` be `true`
+    - New prop `searchOnInputChange`, when it is `true`, `onSearch` will be triggered every time after input changed. The default value is `false`.
+    - New prop `blockDuplicateValueSearch`, when it is `true`, `onSearch` will not be triggerd if input value is same with last time searching.
+    - New prop `blockEmptyValueSearch`, when it is `true`, `onSearch` will not be triggerd if input value is empty.
+    - Rename prop `input` to be `inputProps`.
+- [Core] update `<SearchInput>` styles.
 - [Core] Add centered prop into Modal to make it on top of screen by default (#196)
 - [Core] Shorten width for multiple modal. (#197)
 - [Core] Adjust padding-bottom of modal and column view. (#198)